Add a Makefile
This commit replaces the top-level `clean` script with a `Makefile` that
can be used to centralise project-level build commands.

It also provides a `make serve` task to serve the contents of the
`docs/html` directory locally.

A word of warning: `make clean`, and the prior `clean` script, both
remove `docs/html`. It's not immediately clear how to regenerate it.
Once we know how to do that, the relevant task can be added to the
`Makefile`.