Recreate toolchain from tarball if not present in tree
Download tarball from SiFive and extract to exactly same location, where
it was committed, just before attempting to make any further make rules.

File .generated.linux.gtk.x86_64 is used as an indicator that toolchain
is fine, so download and unpack is done only, if it's not there. This
avoids redownloading on every make call.